Remove Sage namespaced utility classes
