+++ web/app/themes/wordpress-scaffold/templates/partials/site-header.php
@@ -1,6 +1,6 @@
 <?php
 
-use Roots\Sage\Assets;
+use Grrr\Utils\Assets;
 use Grrr\Utils;
 
 ?>
